+++ cmd/proxyguard-client/main.go
@@ -26,7 +26,7 @@ func (cl *ClientLogger) Log(msg string) {
 func main() {
 	fwmark := flag.Int("fwmark", -1, "[Linux only] The fwmark/SO_MARK to use on the underlying TCP socket. -1 is disable.")
 	listen := flag.String("listen", "", "The IP:PORT to listen for UDP traffic.")
-	tcpsp := flag.Int("tcpport", -1, "The PORT to use as the TCP source port. The default is -1, which indicates the same port as the UDP listen. Set this to zero to allocate a freely available port.")
+	tcpsp := flag.Int("tcpport", 0, "The PORT to use as the TCP source port. The default is 0, which means a port chosen by the kernel.")
 	to := flag.String("to", "", "The IP:PORT to which to send the converted TCP traffic to. Specify the server endpoint which also runs Proxyguard.")
 	version := flag.Bool("version", false, "Show version information")
 	pipss := flag.String("peer-ips", "", "Set the IP addresses (separated by commas) to use for the server peer such that DNS resolution does not fail due to timing issues of starting the proxy e.g. on boot, before DNS resolution is ready")