+++ upgrade.go
@@ -1,3 +1,5 @@
 package proxyguard
 
-const UpgradeProto = "wireguard"
+// UpgradeProto is the protocol used in HTTP Upgrade
+// UoTLV/1 is UDP over TCP Length Value version 1
+const UpgradeProto = "UoTLV/1"